// TaskTracker/front/src/features/tasks/enums.ts
export enum Status {
    TODO = 'todo',
    IN_PROGRESS = 'in_progress',
    DONE = 'done',
    IN_REVIEW = 'in_review',
}

export enum Priority {
    LOW = 'low',
    MEDIUM = 'medium',
    HIGH = 'high',
}

export enum TaskType {
    TASK = 'task',
    STORY = 'story',
    EPIC = 'epic',
    BUG = 'bug',
    FEATURE = 'feature',
}